+++ packages/remote-react-components/src/tests/visual/AlertIcon.browser.test.tsx
@@ -1,14 +1,13 @@
 import { testEnvironments } from "@/tests/lib/environments";
 import { test } from "vitest";
-
-const states = ["info", "success", "warning", "danger", "unavailable"] as const;
+import { statusTypes } from "@mittwald/flow-react-components/internal";
 
 test.each(testEnvironments)(
   "AlertIcon states (%s)",
   async ({ testScreenshot, render, components: { AlertIcon, Flex } }) => {
     await render(
       <Flex gap="s">
-        {states.map((status) => (
+        {statusTypes.map((status) => (
           <AlertIcon key={status} status={status} />
         ))}
       </Flex>,